+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Change ReiserFS support to be installed from the AUR.
 - Remove FATX support due to the upstream package frequently breaking.
 - Change GE Proton version from GE-Proton9-18 to GE-Proton9-25.
+- Change Proton Sarek version from Proton-Sarek9-18 to Proton-Sarek9-25-async.
 
 ## 4.2.0 - The Ventoy Update